example_project/requirements.txt
python-dotenv
numpy==1.22.2
scikit-learn==1.0.2